feat(improve_pn_rendering) - refactor CTRM logic to flush push impressions only and remove redundant code SDK-2938
